requirements.txt
pytest==2.8.7
pytest-cov==2.2.1
boto3==1.3.1
botocore==1.4.17
mock==1.3.0
aws_lambda_sample_events==1.0.1
pychef==0.3.0
